Added validations and Authentication Guards.
Removed the firebase deployment script from index.html.

Added a validation for the image when adding a book. So it needs to start with http or/and https, either way if the user is sending a blank string as text (asd/123/blank) it won't work and the validation will trigger.

Added an Authentication Guard for the link for the User Profile and Adding a book. When a user is non-authorized, and is not logged in, by typing the URL (localhost:5173/add-book or localhost:5173/my-profile) they'll get redirected to the login page, or home page, depending on which link they access (adding book -> login // profile -> home).